Rename `tpe` to `traits`.
