+++ src/builtin/matrix.rs
@@ -24,8 +24,8 @@
 // The GLSL Specification, ch 8.6, Matrix Functions.
 
 use num::{ BaseFloat, Zero };
-use mat::tpe::{ GenMat, GenSquareMat };
-use vec::tpe::{ GenVec, GenFloatVec };
+use mat::traits::{ GenMat, GenSquareMat };
+use vec::traits::{ GenVec, GenFloatVec };
 
 /// Multiply matrix `x` by matrix `y` component-wise, i.e., `result[i][j]` is
 /// the scalar product of `x[i][j]` and `y[i][j]`.
