+++ src/builtin/noise.rs
@@ -36,7 +36,7 @@
 #![allow(non_snake_case)]
 
 use num::{ One, Zero };
-use tpe::GenType;
+use traits::GenType;
 use vec::vec::{ Vec2, Vec3, Vec4, vec2, vec3, vec4 };
 use super::common::*;
 use super::geom::dot;
