+++ src/ext/common.rs
@@ -22,7 +22,7 @@
 // THE SOFTWARE.
 
 use num::BaseFloat;
-use tpe::GenFloat;
+use traits::GenFloat;
 use std::num::Float;
 
 /// Returns the reciprocal (inverse) of float number `x`.
